Breakup FhirIO class file.
