+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/healthcare/FhirIOTest.java
@@ -22,6 +22,8 @@ import java.util.Collections;
 import java.util.List;
 import org.apache.beam.sdk.coders.ListCoder;
 import org.apache.beam.sdk.coders.StringUtf8Coder;
+import org.apache.beam.sdk.io.gcp.healthcare.FhirIOSearch.SearchParameter;
+import org.apache.beam.sdk.io.gcp.healthcare.FhirIOSearch.SearchParameterCoder;
 import org.apache.beam.sdk.testing.PAssert;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.transforms.Count;
@@ -42,7 +44,7 @@ public class FhirIOTest {
   @Test
   public void test_FhirIO_failedReads() {
     List<String> badMessageIDs = Arrays.asList("foo", "bar");
-    FhirIO.Read.Result readResult =
+    FhirIORead.Result readResult =
         pipeline.apply(Create.of(badMessageIDs)).apply(FhirIO.readResources());
 
     PCollection<HealthcareIOError<String>> failed = readResult.getFailedReads();
@@ -60,11 +62,11 @@ public class FhirIOTest {
 
   @Test
   public void test_FhirIO_failedSearches() {
-    List<FhirSearchParameter<String>> input =
-        Arrays.asList(FhirSearchParameter.of("resource-type-1", null));
-    FhirIO.Search.Result searchResult =
+    List<SearchParameter<String>> input =
+        Arrays.asList(SearchParameter.of("resource-type-1", null));
+    FhirIOSearch.Result searchResult =
         pipeline
-            .apply(Create.of(input).withCoder(FhirSearchParameterCoder.of(StringUtf8Coder.of())))
+            .apply(Create.of(input).withCoder(SearchParameterCoder.of(StringUtf8Coder.of())))
             .apply(FhirIO.searchResources("bad-store"));
 
     PCollection<HealthcareIOError<String>> failed = searchResult.getFailedSearches();
@@ -81,15 +83,15 @@ public class FhirIOTest {
 
   @Test
   public void test_FhirIO_failedSearchesWithGenericParameters() {
-    List<FhirSearchParameter<List<String>>> input =
-        Arrays.asList(FhirSearchParameter.of("resource-type-1", null));
-    FhirIO.Search.Result searchResult =
+    List<SearchParameter<List<String>>> input =
+        Arrays.asList(SearchParameter.of("resource-type-1", null));
+    FhirIOSearch.Result searchResult =
         pipeline
             .apply(
                 Create.of(input)
-                    .withCoder(FhirSearchParameterCoder.of(ListCoder.of(StringUtf8Coder.of()))))
+                    .withCoder(SearchParameterCoder.of(ListCoder.of(StringUtf8Coder.of()))))
             .apply(
-                (FhirIO.Search<List<String>>)
+                (FhirIOSearch<List<String>>)
                     FhirIO.searchResourcesWithGenericParameters("bad-store"));
 
     PCollection<HealthcareIOError<String>> failed = searchResult.getFailedSearches();
@@ -111,9 +113,9 @@ public class FhirIOTest {
 
     PCollection<String> fhirBundles = pipeline.apply(Create.of(emptyMessages));
 
-    FhirIO.Write.Result writeResult =
+    FhirIOWrite.Result writeResult =
         fhirBundles.apply(
-            FhirIO.Write.executeBundles(
+            FhirIOWrite.executeBundles(
                 "projects/foo/locations/us-central1/datasets/bar/hl7V2Stores/baz"));
 
     PCollection<HealthcareIOError<String>> failedInserts = writeResult.getFailedBodies();
